h-5

+++ kitchen-service/src/main/java/ru/liga/entities/RestauranMenuItemsEntity.java
@@ -29,7 +29,7 @@ public class RestauranMenuItemsEntity {
     @Column(name = "nameitems")
     private String name;
 
-    private double price;
+    private BigDecimal price;
 
     private String image;
